pyproject.toml
[tool.poetry]
name = "inventory-app"
version = "0.1.0"
description = ""
authors = ["250mon <lambki@naver.com>"]
readme = "README.md"
packages = [
    { include = "inventory_app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pandas = "^2.2.3"
pyside6 = "^6.8.0"
sqlalchemy = "^2.0.36"
bcrypt = "^4.2.1"
asyncpg = "^0.30.0"
pyyaml = "^6.0.2"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
qasync = "^0.27.1"
pytest = "^8.3.4"
pytest-qt = "^4.4.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
